+++ src/weapon.c
@@ -23,17 +23,31 @@ bool activeTron = false;
 bool activeChilirun = false;
 
 struct weapon wep_list[N_WEAPONS] = {
-    {wepLightningspeed, 1, "Lightning-speed", "Gain lightning-speed.", ""},
-    {wepFrostwave, 1, "Frost wave", "Slow down your", "opponents."},
-    {wepConfusion, 1, "Confusion", "Inverted controls for", "others."},
-    {wepSharpturn, 3, "Sharp turn", "Perform an ultra sharp", "turn."},
-    {wepTimestep, 1, "Time step", "Jump through time and", "space."},
-    {wepMole, 2, "Mole", "Dig your way out.", ""},
-    {wepWarp, 4, "Warp", "Warp to a random spot", "on the map."},
-    {wepGhost, 1, "Ghost walk", "Transform into ghost", "form."},
-    {wepTron, 1, "Tron-mode", "No more smooth turns.", ""},
-    {wepChilirun, 1, "Chili run", "Hot hot hot!", ""}
+    {wepLightningspeed, addParticlesLightningspeed, 1, "Lightning-speed",
+     "Gain lightning-speed.", ""},
+    {wepFrostwave, addParticlesFrostwave, 1, "Frost wave",
+     "Slow down your", "opponents."},
+    {wepConfusion, addParticlesConfusion, 1, "Confusion",
+     "Inverted controls for", "others."},
+    {wepSharpturn, addParticlesSharpturn, 3, "Sharp turn",
+     "Perform an ultra sharp", "turn."},
+    {wepTimestep, addParticlesVoid, 1, "Time step",
+     "Jump through time and", "space."},
+    {wepMole, addParticlesVoid, 2, "Mole",
+     "Dig your way out.", ""},
+    {wepWarp, addParticlesLightningspeed, 4, "Warp",
+     "Warp to a random spot", "on the map."},
+    {wepGhost, addParticlesGhost, 1, "Ghost walk",
+     "Transform into ghost", "form."},
+    {wepTron, addParticlesTron, 1, "Tron-mode",
+     "No more smooth turns.", ""},
+    {wepChilirun, addParticlesChilirun, 1, "Chili run",
+     "Hot hot hot!", ""}
 };
+
+void addParticlesVoid(struct player *p __attribute__((unused)),
+                      Uint32 delta __attribute__((unused)))
+{}
 
 /**
  * Weapon: Lightning-speed.
@@ -47,6 +61,18 @@ int wepLightningspeed(struct player *p, bool on)
         p->speed = 1.0;
     }
     return DURATION_LIGHTNINGSPEED;
+}
+
+void addParticlesLightningspeed(struct player *p, Uint32 delta)
+{
+    addParticles(delta * 2.5,
+                 p->posx, p->posy,
+                 0, 0.4,
+                 p->dir + M_PI - 0.12, p->dir + M_PI + 0.12,
+                 0.5, 1, 0.0005,
+                 235, 253,
+                 224, 247,
+                 102, 185);
 }
 
 /**
@@ -69,6 +95,23 @@ int wepFrostwave(struct player *p, bool on)
         }
     }
     return DURATION_FROSTWAVE;
+}
+
+void addParticlesFrostwave(struct player *p, Uint32 delta)
+{
+    for (int i = 0; i < nPlayers; ++i) {
+        struct player *target = &players[i];
+        if (target != p && target->alive) {
+            addParticles(delta / 2.0,
+                         target->posx, target->posy,
+                         0, 0.04,
+                         0, 2*M_PI,
+                         1, 1, 0.0005,
+                         28, 28,
+                         212, 212,
+                         228, 228);
+        }
+    }
 }
 
 /**
@@ -97,6 +140,20 @@ int wepConfusion(struct player *p, bool on)
     return DURATION_CONFUSION;
 }
 
+void addParticlesConfusion(struct player *p, Uint32 delta)
+{
+    for (int i = 0; i < nPlayers; ++i) {
+        struct player *target = &players[i];
+        if (target != p && target->alive) {
+            addParticleSpinner(delta, target, 0.03, 20,
+                               0.08, 0.1,
+                               138, 255,
+                               0, 0,
+                               255, 255);
+        }
+    }
+}
+
 /**
  * Weapon: Sharp turn.
  */
@@ -117,15 +174,49 @@ int wepSharpturn(struct player *p, bool on)
     return DURATION_SHARPTURN;
 }
 
+void addParticlesSharpturn(struct player *p, Uint32 delta)
+{
+    addParticles(delta / 2.0,
+                 p->posx, p->posy,
+                 0.1, 0.5,
+                 p->dir + M_PI - 0.05, p->dir + M_PI + 0.05,
+                 1, 1, 0.001,
+                 0, 218,
+                 183, 239,
+                 22, 221);
+}
+
 /**
  * Weapon: Time step.
  */
 int wepTimestep(struct player *p, bool on)
 {
     if (on) {
+        if (particleEffects) {
+            addParticles(50,
+                         p->posx, p->posy,
+                         0.05, 0.2,
+                         0, 2*M_PI,
+                         0.5, 1, 0.0005,
+                         255, 255,
+                         255, 255,
+                         255, 255);
+        }
+
         playSound(SOUND_TIMESTEP, sound);
         p->posx += 90 * cos(p->dir);
         p->posy += 90 * sin(p->dir);
+
+        if (particleEffects) {
+            addParticles(50,
+                         p->posx, p->posy,
+                         0.05, 0.2,
+                         0, 2*M_PI,
+                         0.5, 1, 0.0005,
+                         255, 255,
+                         255, 255,
+                         255, 255);
+        }
     }
     return WEP_NONACTIVE;
 }
@@ -137,6 +228,18 @@ int wepMole(struct player *p, bool on)
 {
     if (on) {
         playSound(SOUND_MOLE, sound);
+
+        if (particleEffects) {
+            SDL_Color color = colors[p->color];
+            addParticles(100,
+                         p->posx, p->posy,
+                         0, 0.1,
+                         p->dir - 0.2, p->dir + 0.2,
+                         0.5, 1, 0.0005,
+                         color.r, color.r,
+                         color.b, color.b,
+                         color.g, color.g);
+        }
 
         double oldposx = p->posx;
         double oldposy = p->posy;
@@ -150,7 +253,7 @@ int wepMole(struct player *p, bool on)
         p->initposy = oldposy;
         p->initdir = olddir + M_PI;
 
-        p->inv_self= DURATION_MOLE;
+        p->inv_self = DURATION_MOLE;
         p->inv_others = DURATION_MOLE;
     }
 
@@ -171,6 +274,17 @@ int wepWarp(struct player *p, bool on)
         rnd = (double)rand()/RAND_MAX;
         p->posy = SPAWN_SPACE_MIN
             + (rnd * (WINDOW_H - 2*SPAWN_SPACE_MIN));
+
+        if (particleEffects) {
+            addParticles(1000,
+                         p->posx, p->posy,
+                         0, 0.5,
+                         0, 2*M_PI,
+                         0.9, 1, 0.0005,
+                         54, 183,
+                         3, 211,
+                         97, 246);
+        }
     }
 
     return WEP_NONACTIVE;
@@ -190,6 +304,18 @@ int wepGhost(struct player *p, bool on)
     return 0;
 }
 
+void addParticlesGhost(struct player *p, Uint32 delta)
+{
+    addParticles(delta * 8,
+                 p->posx, p->posy,
+                 0, 0.2,
+                 0, 2*M_PI,
+                 0.8, 1, 0.0009,
+                 91, 192,
+                 15, 119,
+                 100, 199);
+}
+
 /**
  * Weapon: Tron-mode.
  */
@@ -206,6 +332,30 @@ int wepTron(struct player *p __attribute__ ((unused)), bool on)
     return DURATION_TRON;
 }
 
+void addParticlesTron(struct player *p __attribute__((unused)), Uint32 delta)
+{
+    for (int i = 0; i < nPlayers; ++i) {
+        struct player *target = &players[i];
+        if (target->alive) {
+            addParticles(delta,
+                         target->posx, target->posy,
+                         0, 0.1,
+                         target->dir + M_PI, target->dir + M_PI,
+                         1, 1, 0.0005,
+                         195, 195,
+                         194, 194,
+                         255, 255);
+            addParticles(delta,
+                         target->posx, target->posy,
+                         0, 0.1,
+                         target->dir, target->dir,
+                         1, 1, 0.0005,
+                         195, 195,
+                         194, 194,
+                         255, 255);
+        }
+    }
+}
 
 /**
  * Weapon: Chili run.
@@ -228,4 +378,21 @@ int wepChilirun(struct player *p, bool on)
     }
 
     return DURATION_CHILIRUN;
+}
+
+void addParticlesChilirun(struct player *p, Uint32 delta)
+{
+    for (int i = 0; i < nPlayers; ++i) {
+        struct player *target = &players[i];
+        if (target != p && target->alive) {
+            addParticles(delta * 10,
+                         target->posx, target->posy,
+                         0, 0.5,
+                         0, 2*M_PI,
+                         0.1, 1, 0.001,
+                         223, 255,
+                         31, 220,
+                         26, 220);
+        }
+    }
 }