+++ NEWS.md
@@ -1,5 +1,8 @@
 ## ?.?.? - ??.??.??
 * General:
+  * A particle system has been added. Weapons and player deaths now
+    trigger particle effects. This feature can be turned off in the
+    settings menu.
   * The weapon Switch-aroo has been removed.
 
 ## 0.3.3 - 08.05.20